requirements.txt
tensorflow
tensorflow_datasets
keras
pandas
Jinja2
pydot
fastapi
uvicorn
python-multipart
pydantic
python-dotenv

# pip install --upgrade -r requirements.txt
